backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.2.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
Pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.2
qrcode==7.4.2
reportlab==4.2.5
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
